wip: exploring the right architecture for a plugin system
