+++ cloudvolume/datasource/precomputed/tx.py
@@ -1,29 +1,25 @@
 from functools import partial
-import itertools
-import math
-import multiprocessing as mp
-import concurrent.futures
 import os
-import signal
 
 import numpy as np
 from six.moves import range
 from tqdm import tqdm
 
-from . import lib, chunks
-from .exceptions import AlignmentError, EmptyVolumeException
-from .cacheservice import CacheService
-from .lib import ( 
-  toabs, colorize, red, yellow, 
+from cloudvolume import lib, chunks
+from cloudvolume.exceptions import AlignmentError
+from cloudvolume.cacheservice import CacheService
+from cloudvolume.lib import ( 
   mkdir, clamp, xyzrange, Vec, 
-  Bbox, min2, max2, check_bounds, 
-  jsonify, generate_slices
+  Bbox, min2, max2,
 )
-from .scheduler import schedule_jobs
-from .storage import Storage, SimpleStorage, reset_connection_pools
-from .threaded_queue import ThreadedQueue, DEFAULT_THREADS
-from .volumecutout import VolumeCutout
-from . import sharedmemory as shm
+from cloudvolume.scheduler import schedule_jobs
+from cloudvolume.storage import Storage, SimpleStorage, reset_connection_pools
+from cloudvolume.threaded_queue import ThreadedQueue, DEFAULT_THREADS
+from cloudvolume.volumecutout import VolumeCutout
+
+import cloudvolume.sharedmemory as shm
+
+from .common import fs_lock, parallel_execution, chunknames, shade
 
 NON_ALIGNED_WRITE = yellow(
   """
@@ -53,230 +49,7 @@ NON_ALIGNED_WRITE = yellow(
     Nearest Aligned: {check}
 """)
 
-# Used in sharedmemory to emulate shared memory on 
-# OS X using a file, which has that facility but is 
-# more limited than on Linux.
-fs_lock = mp.Lock()
-
-def parallel_execution(fn, items, parallel, cleanup_shm=None):
-  def cleanup(signum, frame):
-    if cleanup_shm:
-      shm.unlink(cleanup_shm)
-
-  prevsigint = signal.getsignal(signal.SIGINT)
-  prevsigterm = signal.getsignal(signal.SIGTERM)
-
-  signal.signal(signal.SIGINT, cleanup)
-  signal.signal(signal.SIGTERM, cleanup)
-
-  with concurrent.futures.ProcessPoolExecutor(max_workers=parallel) as executor:
-    executor.map(fn, items)
-
-  signal.signal(signal.SIGINT, prevsigint)
-  signal.signal(signal.SIGTERM, prevsigterm)
-
-
-def multi_process_download(cv, bufferbbox, caching, cloudpaths):
-  global fs_lock
-  reset_connection_pools() # otherwise multi-process hangs
-  cv.init_submodules(caching)
-
-  array_like, renderbuffer = shm.bbox2array(cv, bufferbbox, lock=fs_lock)
-  def process(img3d, bbox):
-    shade(renderbuffer, bufferbbox, img3d, bbox)
-  download_multiple(cv, cloudpaths, fn=process)
-  array_like.close()
-
-def multi_process_cutout(vol, requested_bbox, cloudpaths, parallel, 
-  shared_memory_location, output_to_shared_memory):
-  global fs_lock
-
-  cloudpaths_by_process = []
-  length = int(math.ceil(len(cloudpaths) / float(parallel)) or 1)
-  for i in range(0, len(cloudpaths), length):
-    cloudpaths_by_process.append(
-      cloudpaths[i:i+length]
-    )
-
-  spd = partial(multi_process_download, vol, requested_bbox, vol.cache.enabled)
-  parallel_execution(spd, cloudpaths_by_process, parallel, cleanup_shm=shared_memory_location)
-
-  mmap_handle, renderbuffer = shm.bbox2array(vol, requested_bbox, lock=fs_lock, location=shared_memory_location)
-  if not output_to_shared_memory:
-    shm.unlink(vol.shared_memory_id)
-  shm.track_mmap(mmap_handle)
-
-  return mmap_handle, renderbuffer
-
-def cutout(vol, requested_bbox, steps, channel_slice=slice(None), parallel=1, 
-  shared_memory_location=None, output_to_shared_memory=False):
-  """Cutout a requested bounding box from storage and return it as a numpy array."""
-  global fs_lock
-
-  cloudpath_bbox = requested_bbox.expand_to_chunk_size(vol.underlying, offset=vol.voxel_offset)
-  cloudpath_bbox = Bbox.clamp(cloudpath_bbox, vol.bounds)
-  cloudpaths = list(chunknames(cloudpath_bbox, vol.bounds, vol.key, vol.underlying))
-  shape = list(requested_bbox.size3()) + [ vol.num_channels ]
-
-  handle = None
-
-  if parallel == 1:
-    if output_to_shared_memory:
-      array_like, renderbuffer = shm.bbox2array(vol, requested_bbox, 
-        location=shared_memory_location, lock=fs_lock)
-      shm.track_mmap(array_like)
-    else:
-      renderbuffer = np.zeros(shape=shape, dtype=vol.dtype, order='F')
-
-    def process(img3d, bbox):
-      shade(renderbuffer, requested_bbox, img3d, bbox)
-    download_multiple(vol, cloudpaths, fn=process)
-  else:
-    handle, renderbuffer = multi_process_cutout(vol, requested_bbox, cloudpaths, parallel, 
-      shared_memory_location, output_to_shared_memory)
-  
-  renderbuffer = renderbuffer[ ::steps.x, ::steps.y, ::steps.z, channel_slice ]
-  return VolumeCutout.from_volume(vol, renderbuffer, requested_bbox, handle=handle)
-
-def download_single(vol, cloudpath, filename, cache):
-  with SimpleStorage(cloudpath) as stor:
-    content = stor.get_file(filename)
-
-  if cache:
-    with SimpleStorage('file://' + vol.cache.path) as stor:
-      stor.put_file(
-        file_path=filename, 
-        content=(content or b''), 
-        content_type=content_type(vol), 
-        compress=should_compress(vol, iscache=True),
-      )
-
-  bbox = Bbox.from_filename(filename) # possible off by one error w/ exclusive bounds
-  img3d = decode(vol, filename, content)
-  return img3d, bbox
-
-def download_multiple(vol, cloudpaths, fn):
-  locations = vol.cache.compute_data_locations(cloudpaths)
-  cachedir = 'file://' + os.path.join(vol.cache.path, vol.key)
-
-  def process(cloudpath, filename, cache):
-    img3d, bbox = download_single(vol, cloudpath, filename, cache)
-    fn(img3d, bbox)
-
-  local_downloads = ( partial(process, cachedir, filename, False) for filename in locations['local'] )
-  remote_downloads = ( partial(process, vol.layer_cloudpath, filename, vol.cache.enabled) for filename in locations['remote'] )
-
-  downloads = itertools.chain( local_downloads, remote_downloads )
-
-  schedule_jobs(
-    fns=downloads, 
-    concurrency=DEFAULT_THREADS, 
-    progress=('Downloading' if vol.progress else None),
-    total=len(cloudpaths),
-    green=False # not ready for this yet
-  )
-
-def decode(vol, filename, content):
-  """Decode content according to settings in a cloudvolume instance."""
-  bbox = Bbox.from_filename(filename)
-  content_len = len(content) if content is not None else 0
-
-  if not content:
-    if vol.fill_missing:
-      content = ''
-    else:
-      raise EmptyVolumeException(filename)
-
-  shape = list(bbox.size3()) + [ vol.num_channels ]
-
-  try:
-    return chunks.decode(
-      content, 
-      encoding=vol.encoding, 
-      shape=shape, 
-      dtype=vol.dtype, 
-      block_size=vol.compressed_segmentation_block_size,
-    )
-  except Exception as error:
-    print(red('File Read Error: {} bytes, {}, {}, errors: {}'.format(
-        content_len, bbox, filename, error)))
-    raise
-
-def shade(renderbuffer, bufferbbox, img3d, bbox):
-  """Shade a renderbuffer with a downloaded chunk. 
-    The buffer will only be painted in the overlapping
-    region of the content."""
-
-  if not Bbox.intersects(bufferbbox, bbox):
-    return
-
-  spt = max2(bbox.minpt, bufferbbox.minpt)
-  ept = min2(bbox.maxpt, bufferbbox.maxpt)
-
-  ZERO3 = Vec(0,0,0)
-
-  istart = max2(spt - bbox.minpt, ZERO3)
-  iend = min2(ept - bbox.maxpt, ZERO3) + img3d.shape[:3]
-
-  rbox = Bbox(spt, ept) - bufferbbox.minpt
-  if len(img3d.shape) == 3:
-    img3d = img3d[ :, :, :, np.newaxis]
-  
-  renderbuffer[ rbox.to_slices() ] = img3d[ istart.x:iend.x, istart.y:iend.y, istart.z:iend.z, : ]
-
-def content_type(vol):
-  if vol.encoding == 'jpeg':
-    return 'image/jpeg'
-  elif vol.encoding in ('compressed_segmentation', 'fpzip', 'kempressed'):
-    return 'image/x.' + vol.encoding 
-  return 'application/octet-stream'
-
-def should_compress(vol, iscache=False):
-  if iscache and vol.cache.compress != None:
-    return vol.cache.compress
-
-  if vol.compress is None:
-    return 'gzip' if vol.encoding in ('raw', 'compressed_segmentation') else None
-  elif vol.compress == True:
-    return 'gzip'
-  elif vol.compress == False:
-    return None
-  else:
-    return vol.compress
-
-def cdn_cache_control(val):
-  """Translate cdn_cache into a Cache-Control HTTP header."""
-  if val is None:
-    return 'max-age=3600, s-max-age=3600'
-  elif type(val) is str:
-    return val
-  elif type(val) is bool:
-    if val:
-      return 'max-age=3600, s-max-age=3600'
-    else:
-      return 'no-cache'
-  elif type(val) is int:
-    if val < 0:
-      raise ValueError('cdn_cache must be a positive integer, boolean, or string. Got: ' + str(val))
-
-    if val == 0:
-      return 'no-cache'
-    else:
-      return 'max-age={}, s-max-age={}'.format(val, val)
-  else:
-    raise NotImplementedError(type(val) + ' is not a supported cache_control setting.')
-
-def check_grid_aligned(vol, img, offset):
-  """Returns (is_aligned, img bounds Bbox, nearest bbox inflated to grid aligned)"""
-  shape = Vec(*img.shape)[:3]
-  offset = Vec(*offset)[:3]
-  bounds = Bbox( offset, shape + offset)
-  alignment_check = bounds.expand_to_chunk_size(vol.underlying, vol.voxel_offset)
-  alignment_check = Bbox.clamp(alignment_check, vol.bounds)
-  is_aligned = np.all(alignment_check.minpt == bounds.minpt) and np.all(alignment_check.maxpt == bounds.maxpt)
-  return (is_aligned, bounds, alignment_check) 
-
-def upload_image(
+def upload(
     vol, img, offset, 
     parallel=1, 
     manual_shared_memory_id=None, 
@@ -331,7 +104,7 @@ def upload_image(
     # we're throwing them away so safe to write
     img3d.setflags(write=1) 
     shade(img3d, bbox, img, bounds)
-    single_process_upload(
+    threaded_upload_chunks(
       vol, img3d, 
       (( Vec(0,0,0), Vec(*img3d.shape[:3]), bbox.minpt, bbox.maxpt),), 
       n_threads=0,
@@ -353,7 +126,7 @@ def upload_aligned(
   chunk_ranges = list(generate_chunks(vol, img, offset))
 
   if parallel == 1:
-    single_process_upload(
+    threaded_upload_chunks(
       vol, img, chunk_ranges, 
       delete_black_uploads=delete_black_uploads
     )
@@ -426,13 +199,13 @@ def multi_process_upload(
     delta_box = cutout_bbox.clone() - manual_shared_memory_bbox.minpt
     renderbuffer = renderbuffer[ delta_box.to_slices() ]
 
-  single_process_upload(
+  threaded_upload_chunks(
     vol, renderbuffer, chunk_ranges, 
     delete_black_uploads=delete_black_uploads
   )
   array_like.close()
 
-def single_process_upload(
+def threaded_upload_chunks(
     vol, img, chunk_ranges, 
     n_threads=DEFAULT_THREADS,
     delete_black_uploads=False
@@ -510,6 +283,58 @@ def single_process_upload(
     cachestorage.kill_threads()
 
 
+def content_type(vol):
+  if vol.encoding == 'jpeg':
+    return 'image/jpeg'
+  elif vol.encoding in ('compressed_segmentation', 'fpzip', 'kempressed'):
+    return 'image/x.' + vol.encoding 
+  return 'application/octet-stream'
+
+def should_compress(vol, iscache=False):
+  if iscache and vol.cache.compress != None:
+    return vol.cache.compress
+
+  if vol.compress is None:
+    return 'gzip' if vol.encoding in ('raw', 'compressed_segmentation') else None
+  elif vol.compress == True:
+    return 'gzip'
+  elif vol.compress == False:
+    return None
+  else:
+    return vol.compress
+
+def cdn_cache_control(val):
+  """Translate cdn_cache into a Cache-Control HTTP header."""
+  if val is None:
+    return 'max-age=3600, s-max-age=3600'
+  elif type(val) is str:
+    return val
+  elif type(val) is bool:
+    if val:
+      return 'max-age=3600, s-max-age=3600'
+    else:
+      return 'no-cache'
+  elif type(val) is int:
+    if val < 0:
+      raise ValueError('cdn_cache must be a positive integer, boolean, or string. Got: ' + str(val))
+
+    if val == 0:
+      return 'no-cache'
+    else:
+      return 'max-age={}, s-max-age={}'.format(val, val)
+  else:
+    raise NotImplementedError(type(val) + ' is not a supported cache_control setting.')
+
+def check_grid_aligned(vol, img, offset):
+  """Returns (is_aligned, img bounds Bbox, nearest bbox inflated to grid aligned)"""
+  shape = Vec(*img.shape)[:3]
+  offset = Vec(*offset)[:3]
+  bounds = Bbox( offset, shape + offset)
+  alignment_check = bounds.expand_to_chunk_size(vol.underlying, vol.voxel_offset)
+  alignment_check = Bbox.clamp(alignment_check, vol.bounds)
+  is_aligned = np.all(alignment_check.minpt == bounds.minpt) and np.all(alignment_check.maxpt == bounds.maxpt)
+  return (is_aligned, bounds, alignment_check) 
+
 def generate_chunks(vol, img, offset):
   shape = Vec(*img.shape)[:3]
   offset = Vec(*offset)[:3]
@@ -534,15 +359,3 @@ def generate_chunks(vol, img, offset):
     spt = (startpt + bounds.minpt).astype(int)
     ept = (endpt + bounds.minpt).astype(int)
     yield (startpt, endpt, spt, ept)
-
-def chunknames(bbox, volume_bbox, key, chunk_size):
-  for x,y,z in xyzrange( bbox.minpt, bbox.maxpt, chunk_size ):
-    highpt = min2(Vec(x,y,z) + chunk_size, volume_bbox.maxpt)
-    filename = "{}-{}_{}-{}_{}-{}".format(
-      x, highpt.x,
-      y, highpt.y, 
-      z, highpt.z
-    )
-    yield os.path.join(key, filename)
-
-
